Add resource deletion guardrails
Deletion guardrails were implemented by defining new wrapper types
that contain the necessary information to determine whether it's
safe to delete a specific resource.

Shell methods for the store layer were also defined to access those
wrapper types. While the service layer calls those store methods
and determines whether deletion is allowed.

Relevant sentinel errors were also defined in the content package.

Additionally, some new TODOs were added. Mainly replacing List
methods with a ListWithDetails equivalent. Since it would be
a helpful default to be able to see these additional details in
list form.

+++ internal/cms/service/composer.go
@@ -28,6 +28,8 @@ func (s *ComposerService) Get(
 	return composerStore.Get(ctx, id)
 }
 
+// TODO: Replace List with ListWithDetails.
+
 // List returns an array of Composers, sorted by id.
 func (s *ComposerService) List(
 	ctx context.Context,
@@ -57,13 +59,22 @@ func (s *ComposerService) Update(
 }
 
 // Delete attempts to delete a Composer by id.
+//
+// Composers with at least one Piece are protected against deletion.
 func (s *ComposerService) Delete(
 	ctx context.Context,
 	id int,
 ) error {
 	composerStore := store.NewComposerStore(s.pool)
 
-	// TODO: Prevent deletion of Composers with Pieces.
+	composerWithDetails, err := composerStore.GetWithDetails(ctx, id)
+	if err != nil {
+		return err
+	}
+
+	if composerWithDetails.PieceCount > 0 {
+		return content.ErrComposerProtected
+	}
 
 	return composerStore.Delete(ctx, id)
 }
